+++ examples/apps/inline-resizable/Cargo.toml
