+++ examples/apps/inline-resizable/README.md
